Integration of Ruchira Sasanka changes:
 LD_PRELOAD will preload memkind.so as well. - Updates to README. - Fixed a
 bug that would miss allocating very large mallocs in HBW. - Changed default
 memory type to memkind_hbw_preferred. - Minor changes to test. - Removed
 redundant call. - Minor clean up.

+++ examples/autohbw_helper.h
@@ -69,7 +69,7 @@ void disableAutoHBW()
 ////////////////////////////////////////////////////////////////////////////
 // Define the hueristic here.
 ////////////////////////////////////////////////////////////////////////////
-BOOL isAllocInHBW(int size)
+BOOL isAllocInHBW(size_t size)
 {
 
     if (!isAutoHBWEnabled)
@@ -108,7 +108,7 @@ long getLimit(const unsigned long limit, const char lchar)
     //
     if ((limit > 0) && isalpha(lchar)) {
 
-        int mult = 1;              // default multiplier
+        long mult = 1;              // default multiplier
 
         if (toupper(lchar) == 'K') mult = 1024;
         else if (toupper(lchar) == 'M') mult = 1024 * 1024;
